add commom cases
